Enable showing and fix existing warning

+++ Source/PLCrashFrameDWARFUnwind.cpp
@@ -160,7 +160,7 @@ static plframe_error_t plframe_cursor_read_dwarf_unwind_int (task_t task,
         PLCF_ASSERT(fde_info.pc_start < std::numeric_limits<machine_ptr>::max());
 
         /* Initial instructions */
-        err = cfa_state.eval_program(dwarf_section, pc, fde_info.pc_start, &cie_info, &ptr_state, image->byteorder, plcrash_async_mobject_base_address(dwarf_section), cie_info.initial_instructions_offset, cie_info.initial_instructions_length);
+        err = cfa_state.eval_program(dwarf_section, pc, (uint32_t)fde_info.pc_start, &cie_info, &ptr_state, image->byteorder, plcrash_async_mobject_base_address(dwarf_section), cie_info.initial_instructions_offset, cie_info.initial_instructions_length);
         if (err != PLCRASH_ESUCCESS) {
             PLCF_DEBUG("Failed to evaluate CFA at offset of 0x%" PRIx64 ": %d", (uint64_t) fde_info.instructions_offset, err);
             result = PLFRAME_ENOTSUP;
@@ -168,7 +168,7 @@ static plframe_error_t plframe_cursor_read_dwarf_unwind_int (task_t task,
         }
         
         /*  FDE instructions */
-        err = cfa_state.eval_program(dwarf_section, pc, fde_info.pc_start, &cie_info, &ptr_state, image->byteorder, plcrash_async_mobject_base_address(dwarf_section), fde_info.instructions_offset, fde_info.instructions_length);
+        err = cfa_state.eval_program(dwarf_section, pc, (uint32_t)fde_info.pc_start, &cie_info, &ptr_state, image->byteorder, plcrash_async_mobject_base_address(dwarf_section), fde_info.instructions_offset, fde_info.instructions_length);
         if (err != PLCRASH_ESUCCESS) {
             PLCF_DEBUG("Failed to evaluate CFA at offset of 0x%" PRIx64 ": %d", (uint64_t) fde_info.instructions_offset, err);
             result = PLFRAME_ENOTSUP;
@@ -249,7 +249,7 @@ plframe_error_t plframe_cursor_read_dwarf_unwind (task_t task,
         /* Could only happen due to programmer error; eg, an image that doesn't actually match our thread state */
         PLCF_ASSERT(pc <= UINT32_MAX);
 
-        ferr = plframe_cursor_read_dwarf_unwind_int<uint32_t, int32_t>(task, pc, &image->macho_image, current_frame, previous_frame, next_frame);
+        ferr = plframe_cursor_read_dwarf_unwind_int<uint32_t, int32_t>(task, (uint32_t)pc, &image->macho_image, current_frame, previous_frame, next_frame);
     }
     
     plcrash_async_image_list_set_reading(image_list, false);
